feat: pipeline para ingestão e atualização (11)

+++ src/analytics/exec_query.py
@@ -27,7 +27,7 @@ def date_range(start, stop, monthly=False):
     return dates
 
 # executa a query
-def exec_query(table, db_origin, db_target, dt_start, dt_stop, monthly):
+def exec_query(table, db_origin, db_target, dt_start, dt_stop, monthly, mode='append'):
 
     engine_app = sqlalchemy.create_engine(f"sqlite:///../../data/{db_origin}/database.db")
     # grava dados da análise
@@ -38,17 +38,18 @@ def exec_query(table, db_origin, db_target, dt_start, dt_stop, monthly):
 
     for i in tqdm(dates):
 
-        with engine_analytical.connect() as con:
-            try:
-                query_delete = f"DELETE FROM {table} WHERE dtRef = date('{i}', '-1 day')"
-                con.execute(sqlalchemy.text(query_delete))
-                con.commit()
-            except Exception as err:
-                print(err)
+        if mode == 'append':
+            with engine_analytical.connect() as con:
+                try:
+                    query_delete = f"DELETE FROM {table} WHERE dtRef = date('{i}', '-1 day')"
+                    con.execute(sqlalchemy.text(query_delete))
+                    con.commit()
+                except Exception as err:
+                    print(err)
 
         query_format = query.format(date=i)
         df = pd.read_sql(query_format, engine_app)
-        df.to_sql(table, engine_analytical, index=False, if_exists="append")
+        df.to_sql(table, engine_analytical, index=False, if_exists=mode)
 
 
 # %%
@@ -66,6 +67,7 @@ def main():
     parser.add_argument("--start", type=str, default=now) # --start '2024-03-01'
     parser.add_argument("--stop", type=str, default=now)
     parser.add_argument("--monthly", action='store_true')
+    parser.add_argument("--mode", choice=['append','choice'])
     args = parser.parse_args()
 
     exec_query(args.table, args.db_origin, args.db_target, args.start, args.stop, args.monthly)
